Update Level Design 2.unity
I converged the 2 pathways into one single one before reaching the final rectangular play area of the level. This will make the programming aspect of the game simpler and more effective. No need to code for different cases of the level ending, depending on which path is taken by the player.